Rename a few Dart files from camleCase to snake_case
