+++ app/lib/blackboard/blackboard_view.dart
@@ -10,7 +10,7 @@ import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:intl/intl.dart';
 import 'package:sharezone/groups/group_permission.dart';
-import 'package:sharezone/util/api/courseGateway.dart';
+import 'package:sharezone/util/api/course_gateway.dart';
 import 'package:sharezone_common/helper_functions.dart';
 import 'package:sharezone_widgets/theme.dart';
