+++ app/lib/blocs/homework/homework_card_bloc.dart
@@ -11,7 +11,7 @@ import 'dart:async';
 import 'package:bloc_base/bloc_base.dart';
 import 'package:firebase_hausaufgabenheft_logik/firebase_hausaufgabenheft_logik.dart';
 import 'package:rxdart/subjects.dart';
-import 'package:sharezone/util/API.dart';
+import 'package:sharezone/util/api.dart';
 
 class HomeworkCardBloc extends BlocBase {
   final SharezoneGateway api;
