+++ app/lib/comments/comment_view_factory.dart
@@ -8,7 +8,7 @@
 
 import 'package:meta/meta.dart';
 import 'package:sharezone/groups/group_permission.dart';
-import 'package:sharezone/util/api/courseGateway.dart';
+import 'package:sharezone/util/api/course_gateway.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
 import 'comment.dart';
