+++ app/lib/comments/widgets/comment_section_builder.dart
@@ -14,7 +14,7 @@ import 'package:sharezone/comments/comments_bloc.dart';
 import 'package:sharezone/comments/comments_bloc_factory.dart';
 import 'package:sharezone/comments/comments_gateway.dart';
 import 'package:sharezone/comments/widgets/comment_section.dart';
-import 'package:sharezone/util/API.dart';
+import 'package:sharezone/util/api.dart';
 import 'package:user/user.dart';
 
 class CommentSectionBuilder extends StatefulWidget {
