+++ app/lib/groups/src/pages/course/course_details/course_details_bloc.dart
@@ -14,7 +14,7 @@ import 'package:group_domain_models/group_domain_models.dart';
 import 'package:rxdart/subjects.dart';
 import 'package:sharezone/groups/group_permission.dart';
 import 'package:sharezone/groups/src/models/splitted_member_list.dart';
-import 'package:sharezone/util/api/courseGateway.dart';
+import 'package:sharezone/util/api/course_gateway.dart';
 
 class CourseDetailsBloc extends BlocBase {
   final CourseDetailsBlocGateway _gateway;
