+++ app/lib/timetable/timetable_add_event/bloc/timetable_add_event_bloc_dependencies.dart
@@ -9,7 +9,7 @@
 import 'package:meta/meta.dart';
 import 'package:sharezone/markdown/markdown_analytics.dart';
 import 'package:sharezone/timetable/src/models/lesson_length/lesson_length_cache.dart';
-import 'package:sharezone/util/api/timetableGateway.dart';
+import 'package:sharezone/util/api/timetable_gateway.dart';
 
 class TimetableAddEventBlocDependencies {
   final TimetableGateway gateway;
