+++ app/lib/widgets/homework/homework_card.dart
@@ -26,7 +26,7 @@ import 'package:sharezone/pages/homework_page.dart';
 import 'package:sharezone/report/page/report_page.dart';
 import 'package:sharezone/report/report_icon.dart';
 import 'package:sharezone/report/report_item.dart';
-import 'package:sharezone/util/api/connectionsGateway.dart';
+import 'package:sharezone/util/api/connections_gateway.dart';
 import 'package:sharezone/util/navigation_service.dart';
 import 'package:sharezone/widgets/homework/delete_homework.dart';
 import 'package:sharezone_widgets/adaptive_dialog.dart';
