+++ app/test/timetable/mock/mock_timetable_gateway.dart
@@ -10,7 +10,7 @@ import 'package:date/date.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:sharezone/calendrical_events/models/calendrical_event.dart';
 import 'package:sharezone/timetable/src/models/lesson.dart';
-import 'package:sharezone/util/api/timetableGateway.dart';
+import 'package:sharezone/util/api/timetable_gateway.dart';
 import 'package:sharezone_common/references.dart';
 
 class MockTimetableGateway implements TimetableGateway {
